requirements.txt
streamlit
pandas
polars
requests
google-auth
xlsxwriter